Ensure that `test_structural_similarity_dtype` tests np.int32

+++ tests/skimage/metrics/test_structural_similarity.py
@@ -89,8 +89,8 @@ def test_structural_similarity_dtype(dtype):
     Y = np.random.rand(N, N)
     if np.dtype(dtype).kind in 'iub':
         data_range = 255.0
-        X = (X * 255).astype(np.uint8)
-        Y = (Y * 255).astype(np.uint8)
+        X = (X * 255).astype(dtype)
+        Y = (Y * 255).astype(dtype)
     else:
         data_range = 1.0
         X = X.astype(dtype, copy=False)
